requirements.txt

psycopg2-binary
django
django-widget-tweaks
django-db-file-storage
pandas
python-dotenv
requests
wsgiref